Review comment: remove useless cast.

+++ community/community-it/record-storage-engine-it/src/test/java/org/neo4j/internal/batchimport/store/BatchingNeoStoresTest.java
@@ -195,7 +195,7 @@ class BatchingNeoStoresTest
                     RecordStore<AbstractBaseRecord> store = stores.getNeoStores().getRecordStore( type );
                     if ( type == typeToTest )
                     {
-                        assertThat( (int) store.getHighId() ).as( store.toString() ).isGreaterThan( store.getNumberOfReservedLowIds() );
+                        assertThat( store.getHighId() ).as( store.toString() ).isGreaterThan( store.getNumberOfReservedLowIds() );
                     }
                     else
                     {
